# back/src/swagger.json
{
    "swagger": "2.0",
    "info": {
      "version": "1.0.0",
      "title": "MaxAirain Express API with Swagger",
      "description": "This is a simple CRUD API application made for MaxAirain Company, with Express and documented with Swagger",
      "license": {
        "name": "MIT",
        "url": "https://opensource.org/licenses/MIT"
      }
    },
    "host": "localhost:3000",
    "basePath": "/",
    "tags": [
      {
        "name": "Users",
        "description": "API for users in the system"
      }
    ],
    "schemes": ["http"],
    "consumes": ["application/json"],
    "produces": ["application/json"],
    "paths": {
          "/users": {
            "get": {
              "tags": ["Users"],
              "summary": "Get all users in system",
              "responses": {
                "200": {
                  "description": "OK"
                }
              }
            }
          },
          "/users/{userId}": {
            "parameters": [
                {
                  "name": "userId",
                  "in": "path",
                  "required": true,
                  "description": "ID of user that we want to find",
                  "type": "integer"
                }
              ],
            "get": {
                "tags": ["Users"],
                "summary": "Get user in system by his ID",
                "responses": {
                  "200": {
                    "description": "OK"
                  }
                }
              }
          },
          "/{adminId}/create-user": {
            "parameters": [
                {
                  "name": "adminId",
                  "in": "path",
                  "required": true,
                  "description": "ID of admin",
                  "type": "integer"
                }
              ],
            "post": {
                "tags": ["User"],
                "summary": "Created user in system",
                "responses": {
                  "200": {
                    "description": "OK"
                  }
                }
              }
          },
          "/login": {
            "post": {
                "tags": ["User"],
                "summary": "Login for user",
                "responses": {
                  "200": {
                    "description": "OK"
                  }
                }
              }
          },
          "/{adminId}/users/{userId}": {
            "parameters": [
                {
                  "name": "adminId",
                  "in": "path",
                  "required": true,
                  "description": "ID of user that we want to find",
                  "type": "integer"
                },
                {
                    "name": "userId",
                    "in": "path",
                    "required": true,
                    "description": "ID of the admin",
                    "type": "integer"
                }
              ],
            "delete": {
                "tags": ["User"],
                "summary": "Delete user in system by his ID",
                "responses": {
                  "200": {
                    "description": "OK"
                  }
                }
              }
          },
          "/items": {
            "get": {
              "tags": ["Items"],
              "summary": "Get all items in system",
              "responses": {
                "200": {
                  "description": "OK"
                }
              }
            }
          },
          "/chooseItems/{userId}": {
            "parameters": [
              {
                "name": "userId",
                "in": "path",
                "required": true,
                "description": "ID of user that we want to find",
                "type": "integer"
              }
            ],
            "post": {
                "tags": ["Items"],
                "summary": "Get user in system by his ID",
                "responses": {
                  "200": {
                    "description": "OK"
                  }
                }
            }
          },
          "/items/{userId}" : {
            "parameters": [
              {
                "name": "userId",
                "in": "path",
                "required": true,
                "description": "ID of user that we want to find",
                "type": "integer"
              }
            ],
            "get": {
              "tags": ["Items"],
              "summary": "Get user in system by his ID",
              "responses": {
                "200": {
                  "description": "OK"
                }
              }
            }
          },
          "/{adminId}/create-item": {
            "parameters": [
              {
                "name": "adminId",
                "in": "path",
                "required": true,
                "description": "ID of admin that we want to find",
                "type": "integer"
              }
            ],
            "post": {
                "tags": ["Item"],
                "summary": "Get user in system by his ID",
                "responses": {
                  "200": {
                    "description": "OK"
                  }
                }
              }
          },
          "/update/{itemId}": {
            "parameters": [
              {
                "name": "itemId",
                "in": "path",
                "required": true,
                "description": "ID of item that we want to find",
                "type": "integer"
              }
            ],
            "put": {
                "tags": ["Item"],
                "summary": "Get user in system by his ID",
                "responses": {
                  "200": {
                    "description": "OK"
                  }
                }
              }
          },
          "/{adminId}/items/{itemId}": {
            "parameters": [
              {
                "name": "adminId",
                "in": "path",
                "required": true,
                "description": "ID of admin that we want to find",
                "type": "integer"
              },
              {
                "name": "itemId",
                "in": "path",
                "required": true,
                "description": "ID of item that we want to find",
                "type": "integer"
              }
            ],
            "delete": {
              "tags": ["Item"],
              "summary": "Get user in system by his ID",
              "responses": {
                "200": {
                  "description": "OK"
                }
              }
            }
          }
        
    },
    "definitions": {
      "User": {
        "required": ["first_name", "last_name", "role","picture_face"],
        "properties": {
          "_id": {
            "type": "integer",
            "uniqueItems": true
          },
          "first_name": {
            "type": "string"
          },
          "last_name": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "post": {
            "type": "string"
          },
          "picture_face": {
            "type": "array"
          }
        }
      },
      "Users": {
        "type": "array",
        "$ref": "#/definitions/User"
      },
      "Item": {
        "required": ["label", "quantity", "id"],
        "properties": {
          "_id": {
            "type": "integer",
            "uniqueItems": true
          },
          "quantity": {
            "type": "integer"
          },
          "label": {
            "type": "string"
          }
        }
      },
      "Items": {
        "type": "array",
        "$ref": "#/definitions/Item"
      }
    }
}

  